requirements.txt
django-environ
tweepy
requests
pandas
beautifulsoup4